requirements.txt
Flask==2.3.2
PyJWT==2.8.0
PyJWT==2.8.0
Requests==2.31.0
